requirements.txt
Django

gunicorn

pytz
django-embed-video
pillow